Recreating the source code used by "bad bootloader"
The "bad bootloader" found its way into "x16-smc-r45.1-bootloader.hex",
which then found its way into all PR boards up to ~900. This is a
recreation of the source code which will generate the machine code for this
"bad bootloader".

This version behaves similar to bootloader v2: It programs all pages except
for the first one. Bootloader v2 stores the first page in RAM and performs
a watchdog reset. After reset, it programs the first page. The bad
bootloader, however, enters an infinite loop before activating the watchdog
peripheral. This causes that, if the user disconnects power and reconnects
power, the first page gets programmed with whatever is in RAM after a
coldstart, which is undefined. Causing SMC to become bricked. If, however,
the SMC is restarted with an external reset input (set pin 10 to GND), the
contents of SRAM will be preserved, and the first page is programmed
correctly.